add AbstractAllowedFieldValidator
