+++ src/main/java/com/nephest/battlenet/sc2/model/validation/SortParameterAllowedFieldValidator.java
@@ -4,36 +4,15 @@
 package com.nephest.battlenet.sc2.model.validation;
 
 import com.nephest.battlenet.sc2.model.web.SortParameter;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
 
 public class SortParameterAllowedFieldValidator
-implements ConstraintValidator<AllowedField, SortParameter>
+extends AbstractAllowedFieldValidator<SortParameter>
 {
 
-    private Set<String> allowedFields;
-
     @Override
-    public void initialize(AllowedField constraintAnnotation)
+    public String getField(SortParameter value)
     {
-        this.allowedFields = new HashSet<>(Arrays.asList(constraintAnnotation.value()));
-    }
-
-    @Override
-    public boolean isValid(SortParameter value, ConstraintValidatorContext context)
-    {
-        if(value == null || allowedFields.contains(value.field())) return true;
-
-        context.disableDefaultConstraintViolation();
-        context.buildConstraintViolationWithTemplate
-        (
-            String.format("Invalid sort field '%s'. Allowed fields are: %s",
-            value.field(), String.join(", ", allowedFields))
-        ).addConstraintViolation();
-        return false;
+        return value == null ? null : value.field();
     }
 
 }
